refactor(chat): extract pure helpers and builtin runners from ChatView
view.ts had regrown to 2,374 lines. Move the pure-function tail into
chat/context-usage.ts and agents/summary.ts, isContinuationToast into
chat/continuation-state.ts, and the /compact ... /fork implementations
into chat/builtin-runners.ts (deps-closure pattern, mirroring
SubagentDispatch). Unify subagent-tracker's duplicate tool-name set
with SUBAGENT_TOOLS now that the circular-import reason is gone, and
drop the summarizePrompt re-export in favor of direct imports.
No behavior change; view.ts is down to 1,917 lines.
